+++ changelog.md
@@ -6,6 +6,8 @@
 This version is not released yet and is under active development.
 ```
 
+* \[brew,cask\] Fix execution of `sync` command.
+
 ## {gh}`4.11.0 (2022-04-03) <compare/v4.10.0...v4.11.0>`
 
 - \[brew,cask\] Do not let homebrew auto-update on other commands. Refs {issue}`36`.